+++ photutils/profiles/__init__.py
@@ -1,0 +1,8 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""
+This subpackage contains tools for generating radial profiles.
+"""
+
+from .core import *  # noqa: F401, F403
+from .curve_of_growth import *  # noqa: F401, F403
+from .radial_profile import *  # noqa: F401, F403